refactor: :zap: changed folder structure for fsd style
