+++ src/pages/SignUp/ui/index.ts
@@ -1,0 +1,1 @@
+export * from './SignUp';
